+++ tools/generate_profile_tables.py
@@ -47,6 +47,7 @@ PROFILE_ORDER = [
     "melsec:iq-r:rj71en71",
     "melsec:iq-l",
     "melsec:mx-r",
+    "melsec:mx-r:rj71en71",
     "melsec:mx-f",
     "melsec:iq-f",
     "melsec:qcpu",